requirements.txt
attrs==20.3.0
click==7.1.2
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
iniconfig==1.1.1
packaging==20.9
pkg-resources==0.0.0
pluggy==0.13.1
py==1.10.0
pycryptodome==3.10.1
pydantic==1.8.1
pyparsing==2.4.7
pytest==6.2.3
SQLAlchemy==1.4.11
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
uvicorn==0.13.4
